src/tryit.jade: Remove cookie
Many pump nodes are down. If someone clicks Try It and gets sent to an offline node they might want to go back and Try It again. Not realizing that a cookie has been sent they keep going back to the same place over and over. Head, meet wall. Hopefully removing the cookie means they can potentially be sent to a different node.